created oauthlogin.js screen to avoid merge conflict
